Move trie modules to /tries and trie libs to /lib/tries
