+++ lib/tries/serializeTrie.js
