+++ tries/private.complete.js
@@ -1,0 +1,5 @@
+"use strict";
+
+const parseTrie = require("../lib/tries/parseTrie");
+
+module.exports = parseTrie(require("../build/lists/private.complete.json").trie);
